+++ packages/shared/src/index.ts
@@ -1,3 +1,4 @@
 export * from "./date";
 export * from "./tasks";
 export * from "./icons";
+export * from "./templates";